// apps/nextjs-app/src/lib/api/models/ApiErrorResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ValidationErrorDetail } from './ValidationErrorDetail';
export type ApiErrorResponse = {
    code?: string;
    message?: string;
    details?: Array<ValidationErrorDetail>;
};